feat: implement TOWERS and FUTOSHIKI puzzles across all targets
- Procedural solvers & canonical datasets: generated 100 verified unique puzzles per tier (easy, medium, hard) for both TOWERS and FUTOSHIKI with D4 dihedral symmetry invariance.
- Python BasePuzzle plugins: added app/puzzles/towers.py and app/puzzles/futoshiki.py with ASCII receipt formatting, canonical descriptions, and 576-dot thermal rasterizers.
- ESP32 & Arduino firmware: implemented TowersGen and FutoshikiGen with PROGMEM flash arrays, nibble packing, and zero-latency generation; synchronized to MorningPuzzles/.
- Bitmask & config expansion: expanded offline puzzle bitmask to 18 games (0x3FFFF) in OfflineConfigManager, PortalHtml, and portal.html captive portal.
- Web Simulator parity: added bit-level physical thermal emulation and UI controls for TOWERS and FUTOSHIKI in simulator/receipt_simulator.html.
- Test suites & visual consistency: added test_towers_deduction.py and test_futoshiki_deduction.py; achieved 100.00% pixel parity across C++, Python, and Simulator.

+++ esp32-firmware/test/test_offline_generators.cpp
@@ -94,6 +94,14 @@ public:
 #include "../src/generators/LadderGen.h"
 #include "../src/generators/LadderGen.cpp"
 
+#include "../src/generators/TowersDataset.h"
+#include "../src/generators/TowersGen.h"
+#include "../src/generators/TowersGen.cpp"
+
+#include "../src/generators/FutoshikiDataset.h"
+#include "../src/generators/FutoshikiGen.h"
+#include "../src/generators/FutoshikiGen.cpp"
+
 #include "../src/generators/OfflinePuzzleComposer.h"
 #include "../src/generators/OfflinePuzzleComposer.cpp"
 
@@ -146,8 +154,8 @@ int main() {
     OfflineConfigManager config;
     config.begin();
     // Enable 7 games: Sudoku, Queens, Mines, Bridges, Tango, Lights, Wheel
-    uint16_t mask = (1 << PUZZLE_SUDOKU) | (1 << PUZZLE_QUEENS) | (1 << PUZZLE_MINES) |
-                    (1 << PUZZLE_BRIDGES) | (1 << PUZZLE_TANGO) | (1 << PUZZLE_LIGHTS) | (1 << PUZZLE_WHEEL);
+    uint32_t mask = (1UL << PUZZLE_SUDOKU) | (1UL << PUZZLE_QUEENS) | (1UL << PUZZLE_MINES) |
+                    (1UL << PUZZLE_BRIDGES) | (1UL << PUZZLE_TANGO) | (1UL << PUZZLE_LIGHTS) | (1UL << PUZZLE_WHEEL);
     config.setGameMask(mask);
     config.setPuzzleCount(4);
     config.setPuzzleGrade(GRADE_ESCALATING);
